refactor: follow fastapi best practices
