--- app/src/routers/handlers/http_error.py
@@ -1,6 +1,0 @@
-from fastapi import HTTPException
-from starlette.requests import Request
-from starlette.responses import JSONResponse
-
-async def http_error_handler(_: Request, exc: HTTPException) -> JSONResponse:
-  return JSONResponse({"errors": [exc.detail]}, status_code=exc.status_code)
