+++ app/src/user/router.py
@@ -1,10 +1,9 @@
 from typing import List
 from fastapi import Depends, APIRouter
 from sqlalchemy.orm import Session
-
 from app.src.dependencies import get_db
-from app.src.domain.user.schemas import User, UserCreate, UserList
-from app.src.domain.user.services import UserService
+from app.src.user.schemas import User, UserCreate, UserList
+from app.src.user.services import UserService
 
 router = APIRouter(tags=["users"])
